Create a properly shrinking grid UI
The grid now properly shrinks instead of
overflowing in the y direction. The key points to
this fix is: restricting the body height well below
100vh, having #container become a flex container
in the y direction, specifiying a height property
value for #container to allow the row divs to
shrink through flex: 1 1 auto, and specifiying
the children of container to have no minimum
width or height.

Restricting the size of body was necessary as
some of the viewport space is being taken up
by an h1 heading. If it's removed,
body{height:100} does indeed work.

Having #container become a flex container in
the y direction means that it will grow and shrink
in the y direction. Also specifiying a height
for it allows #container to shrink child
elements until its height becomes what it
should be.

Setting min-width:0 and min-height: 0 for the
\#container's children allow them to shrink
properly, and prevent overflow.

The downside of this approach is that the squares
will no longer remain dimensionally square, due
to shrinking in the y direction. They can end up
looking more like rectangles.